refactor: break apart plugin into smaller Lua files
